requirements.txt
spacy==3.5.2
tqdm==4.64.1
pytesseract==0.3.10
pillow==9.2.0
flask==1.1.2
flask-mysql==1.5.2
flask-bcrypt==1.0.1
flask-cors==3.0.10
flask-socketio==5.3.4
werkzeug==2.0.3